animation modularized
